refactor configuratt into package. Added _require and _contingent keywords to make non-configured sections optional (i.e. when a cab is defined)
