+++ tests/scabha_tests/testconf.yaml
@@ -29,6 +29,30 @@ scrubbed2:
     - x.a.*
     - x.y
 
+requirements:
+  _include: (scabha_tests)test_include2.yaml
+  x:
+    a:
+      _requires: scrubbed.foo.x
+      b: 2
+
+requirements2:
+  _include: (scabha_tests)test_include2.yaml
+  x:
+    zz:
+      _contingent: scrubbed.foo.x
+      b: 2
+    yy:
+      _contingent: scrubbed.foo.xxx
+      b: 2
+
+requirements3:
+  _include: (scabha_tests)test_include2.yaml
+  x:
+    zz:
+      _requires: scrubbed.foo.xxx
+      b: 2
+
 
 # this is meant to inject into x.y1, the name should also show up in x.y2 via its _use of x.y1
 x:
